feat(model): restrict supported dtypes to float32 and float64
Add a model_dtype helper that validates float32/float64 and clamps the
torch default for model-free utilities. MultiStateJointModel.to rejects
unsupported dtypes before mutating the model, and construction validates
the parameter dtype. The C++ survival builders take a float64 flag and
materialize time arrays directly in the requested dtype, so _surv no
longer casts on the Python side.

+++ tests/test_plot.py
@@ -16,7 +16,7 @@ matplotlib.use("Agg")
 def test_params_history():
     model, data = _model(max_iter=2, tol=0.0), _data()
     model.n_subsample = 1
-    model.to(torch.bfloat16)
+    model.to(torch.float64)
     model.fit(data)
     figure, axes = plot_params_history(model)
     assert len(axes) >= 1
